refactor(config): move UniverseConfig to its own config/universe.py
- Relocate UniverseConfig verbatim from config/system.py to config/universe.py
  (one-domain-per-file convention); pure mechanical move, zero behavior change
- system.py keeps Environment/LogLevel/SystemSettings; drop now-unused Field import
- Barrel re-exports UniverseConfig from .universe (__all__ unchanged)
- Repoint consumers: itrader_config.py, config_router.py, live_trading_system.py (2 sites)
